requirements.txt
opencv-contrib-python-headless
tensorflow-cpu
streamlit
numpy
pandas
pillow
keras
pathlib
matplotlib 
python-resize-image
protobuf>=3.20.0,<4.0.0